+++ src/plugins/pairing_acceleration/pairing_acceleration.py
@@ -10,6 +10,7 @@ from plugins.hookspec import hookimpl
 from plugins.migration import PluginMigrationManager
 from plugins.pairing_acceleration import PLUGIN_NAME, migrations
 from plugins.pairing_acceleration.pairing_variations import (
+    CustomAccelerationSwissVariation,
     HaleySwissVariation,
     HaleySoftSwissVariation,
     ProgressiveSwissVariation,
@@ -60,6 +61,7 @@ class PairingAccelerationPlugin(Plugin):
             HaleySwissVariation,
             HaleySoftSwissVariation,
             ProgressiveSwissVariation,
+            CustomAccelerationSwissVariation,
         ]
 
     @hookimpl